Fix SvgText

+++ src/MudBlazor/Components/Chart/Svg/SvgText.cs
@@ -2,8 +2,8 @@ namespace MudBlazor.Charts.SVG.Models
 {
     public class SvgText
     {
-        public decimal X { get; set; }
-        public decimal Y { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
         public string Value { get; set; }
     }
 }
